User Management System Integration
- userSystem.py was set up to run its own flask app separate from our primary one, which prevented it from functioning. Moved all the code into app.py as a surefire solution. (potentially could've altered userSystem.py and imported it in order to improve organization, meh)

- Updated requirements.txt with the new imports

- Removed email and password reset functionality due to time constraints.

- Adjusted app.py routes to line up with new user system routes.

- Removed old temporary login method

- Adjusted return values to include a readable success field alongside the message
