+++ .gitignore
@@ -92,3 +92,4 @@ cython_debug/
 
 # Ruff stuff:
 .ruff_cache/
+/migrations
