+++ requirements.txt
@@ -1,11 +1,61 @@
+alembic==1.15.2
+bcrypt==4.3.0
 blinker==1.8.2
+certifi==2025.1.31
+chardet==5.2.0
+charset-normalizer==3.4.1
 click==8.1.8
 colorama==0.4.6
-flask==3.0.3
+contourpy==1.3.1
+cssselect2==0.7.0
+cycler==0.12.1
+docker==7.1.0
+dotenv==0.9.9
+Flask==3.0.3
 Flask-Cors==5.0.0
-importlib-metadata==8.5.0
+Flask-Login==0.6.3
+Flask-Mail==0.10.0
+Flask-Migrate==4.1.0
+Flask-SQLAlchemy==3.1.1
+fonttools==4.55.8
+greenlet==3.2.0
+idna==3.10
+importlib_metadata==8.5.0
 itsdangerous==2.2.0
-jinja2==3.1.6
-MarkupSafe==2.1.5
-werkzeug==3.0.6
+Jinja2==3.1.6
+kiwisolver==1.4.8
+lxml==5.3.0
+Mako==1.3.10
+MarkupSafe==3.0.2
+matplotlib==3.10.0
+mysql-connector-python==9.2.0
+mysql-connector-repackaged==0.3.1
+mysqlclient==2.2.7
+netifaces==0.11.0
+numpy==2.2.2
+opencv-python==4.11.0.86
+packaging==24.2
+passlib==1.7.4
+pdf2image==1.17.0
+pillow==11.0.0
+psutil==6.1.0
+PyMuPDF==1.25.1
+PyMySQL==1.1.1
+pyparsing==3.2.1
+python-dateutil==2.9.0.post0
+python-dotenv==1.1.0
+pywin32==310
+reportlab==4.2.5
+requests==2.32.3
+scipy==1.15.2
+six==1.17.0
+SQLAlchemy==2.0.40
+svglib==1.5.1
+tabulate==0.9.0
+tinycss2==1.4.0
+typing_extensions==4.13.2
+urllib3==2.4.0
+webencodings==0.5.1
+Werkzeug==3.0.6
+wsl==0.12
 zipp==3.20.2
